pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "RepoA"
version = "2.1.0"
description = "A comprehensive, modular agent system with Memory management, KV cache optimization, and cost-efficient LLM integration"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "VCoder4646", email = "vasanthwork0475@gmail.com"}
]
keywords = ["agent", "llm", "ai", "memory", "tools", "mcp", "cache"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]
llm = [
    "openai>=1.0.0",
    "anthropic>=0.8.0",
]
data = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
]
database = [
    "sqlalchemy>=2.0.0",
]
async = [
    "asyncio>=3.4.3",
    "aiohttp>=3.9.0",
]
all = [
    "RepoA[dev,llm,data,database,async]",
]

[project.urls]
Homepage = "https://github.com/VCoder4646/RepoA"
Documentation = "https://github.com/VCoder4646/RepoA/blob/main/README.md"
Repository = "https://github.com/VCoder4646/RepoA"
"Bug Reports" = "https://github.com/VCoder4646/RepoA/issues"

[tool.setuptools]
packages = ["repoa", "repoa.core", "repoa.tools", "repoa.config", "repoa.cli"]

[tool.setuptools.package-data]
repoa = ["py.typed"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=repoa --cov-report=html --cov-report=term-missing"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true
